Fix compile errors on Linux
One thing that worked on Windows but not here is related to
the callable ndobject.f("funcname", params...). On Windows,
because it knew the information later when it was instantiated,
allowed callable to be just a forward declaration earlier.
Not so in gcc-land, it requires that the information be there
at the declaration point, even though it knows the information
later when the instantiation is happening. Will need to figure
out a way around this, as the current workaround is unpleasant.