Fix overflow test failures on Linux/Mac
This is a longstanding bug, caused by there being no apparent
way to tell the compiler only that the FP status registers are
important, and not to reorder FP operations and FP status register
checks. The workaround is to explicitly compare against the max
and min float values when casting from double.