Rename fixed_dim type to cfixed_dim
This is a change to make the dynd type system better match datashape.
In datashape, "N * int32" is syntactic sugar for "fixed[N] * int32",
but the previous dynd fixed_dim type doesn't quite match because it
also locks down the stride. To make this match, we rename fixed_dim
to cfixed_dim, following the struct/cstruct idea, then create a new
fixed_dim type which does not lock down the stride.